move player to pkg
